Splitting math library in smaller headers
