+++ magnolia/src/math/generic.hpp
@@ -1,0 +1,10 @@
+#pragma once
+
+#include "glm/glm.hpp"
+#include "glm/gtc/type_ptr.hpp"
+#include "glm/gtx/matrix_decompose.hpp"
+#include "glm/gtx/quaternion.hpp"
+
+// #include "glm/gtx/euler_angles.hpp"
+// #include "glm/gtx/string_cast.hpp"
+// #include "glm/gtx/transform.hpp"
