+++ magnolia/src/math/mat.hpp
@@ -1,0 +1,3 @@
+#pragma once
+
+#include "glm/mat4x4.hpp"
