+++ magnolia/src/math/quat.hpp
@@ -1,0 +1,3 @@
+#pragma once
+
+#include "glm/gtx/quaternion.hpp"
