+++ magnolia/src/math/trigonometric.hpp
@@ -1,0 +1,4 @@
+#pragma once
+
+#include "glm/ext/quaternion_trigonometric.hpp"
+#include "glm/trigonometric.hpp"
