+++ magnolia/src/math/vec.hpp
@@ -1,0 +1,5 @@
+#pragma once
+
+#include "glm/vec2.hpp"
+#include "glm/vec3.hpp"
+#include "glm/vec4.hpp"
