+++ magnolia/src/renderer/render_graph.hpp
@@ -2,6 +2,7 @@
 
 #include <map>
 
+#include "math/vec.hpp"
 #include "renderer/renderer_image.hpp"
 
 namespace mag
