+++ magnolia/src/renderer/shader.cpp
@@ -5,6 +5,7 @@
 #include "core/application.hpp"
 #include "core/assert.hpp"
 #include "core/logger.hpp"
+#include "math/generic.hpp"
 #include "renderer/context.hpp"
 #include "renderer/descriptors.hpp"
 #include "renderer/frame.hpp"
