+++ sprout_editor/src/editor_scene.hpp
@@ -1,6 +1,7 @@
 #pragma once
 
 #include "math/types.hpp"
+#include "math/vec.hpp"
 #include "scene/scene.hpp"
 
 namespace mag
