+++ sprout_editor/src/panels/camera_panel.cpp
@@ -4,6 +4,8 @@
 #include "core/types.hpp"
 #include "icon_font_cpp/IconsFontAwesome6.h"
 #include "imgui.h"
+#include "math/generic.hpp"
+#include "math/type_definitions.hpp"
 #include "math/types.hpp"
 
 namespace sprout
